+++ main/pwm_control.h
@@ -2,6 +2,6 @@
 #define PWM_CONTROL_H
 
 void pwm_init(void);
-void pwm_change_duty(const uint16_t comp_value);
+void pwm_change_duty(const uint32_t comp_value);
 
 #endif
